New content and git.sh will now create bare repositories

+++ lib/git.sh
@@ -13,6 +13,28 @@ export HOME=$tmp_home
 
 ################################################################################
 git_cleanup_temp() {
+  # Push all branches back to origin:
+  for branch in .git/refs/heads/*; do
+    branch=$(basename "$branch")
+    git checkout "$branch"
+    git push origin "$branch"
+  done
+
+  # Push all tags too:
+  git push --tags
+
+  # Remove the non-bare repo:
+  name=$(basename "$(pwd)")
+  cd ..
+
+  if [ "$(basename "$(pwd)")" != "repos" ]; then
+    >&2 echo "BUG: tried to escape the repos dir!"
+    exit 1
+  fi
+
+  rm -rf "$name"
+
+  # Remove the temporary home directory:
   if [ -e "$tmp_home" ]; then
     rm -r "$tmp_home"
   fi
@@ -27,15 +49,21 @@ git_set_author() {
 ################################################################################
 git_init() {
   directory=$1; shift
+  bare="$directory".git
 
-  if [ -e "$directory" ]; then
-    >&2 echo "repository already exists: $directory"
+  if [ -e "$bare" ] || [ -e "$directory" ]; then
+    >&2 echo "repository already exists: $bare"
     exit 1
   fi
 
-  mkdir -p "$directory"
+  mkdir -p "$bare"
+  cd "$bare"
+  git init --bare
+
+  cd "$(dirname "$directory")"
+  git clone "$bare" "$directory"
   cd "$directory"
-  git init
+  git checkout -b master
 }
 
 ################################################################################
